Standardize Project Structure
This commit changes the project to follow common golang and cobra app layout conventions. The `main.go` entry point is moved to the root of the project and all cobra command files are moved to named folders under `cmd`.

This also adds build scripts for Windows and *nix that will build the project with a one word command. These only build the project and ensure that the output is named `pdk`. It is simple on purpose, as `goreleaser` handles the actual compiling and releasing of the application. This is enough for now while the project is new, we can revisit these shell scripts when building the project becomes more complex.