Puppet Content Templates package
Creates the pct golang package which allows users to list and deploy Puppet Content Templates. There is much more to implement, but this is the start.

It can list and format the display of the PCTs installed on a local filesystem. This shows PCT details like id and names and provides autocomplete functionality for the shell commands.

It can deploy a selected template to a target file path based on the `pct-config.yml` configuration file inside the PCT and from the `~/.pdk/pct.yml` user level configuration file.