Implement pdk new
Removes all subcomands for `pdk new` and adapts `pdk new` to use the PCT golang package.

The unit tests for `pdk new` pass but do not actually test any PCT functionality. Tests for PCT functionality are inside the PCT package. A future work item is to either adapt a unit test that can accurately exercise the command line here, or add that as an acceptance test instead.