Set deployment artifact binary name and version
Sets the binary name to `pct` to reflect that this is currently an experiment of the PCT system. This also ensures that use of this binary does not conflict with the shipped PDK command name until we start shipping the final form.

Updates goreleaser config file to set the version, short commit hash, and commit date inside the binary as its compiled. This enables accurate deployment information in `pdk --version` as well as for use inside templates.