(GH-20) Set default 'templatepath' to "$INSTALL_ROOT/templates"
This commit modifies `new.go` to set the default template path to
be the `$INSTALL_ROOT/templates` directory, enabling a "ready to go"
experience for the PCT LA. When a user runs `pct new --list` they
will see the default templates we have elected to ship with the package.

Also added a warning log message in the `pct#Get` function to alert
the user if no templates could be found.

The build scripts can now create a build for local development
purposes or a production ready snapshot using the arguments, "build"
or "snapshot", respectively. The scripts will also handle fetching
the default content templates from `puppetlabs/baker-round`.