(GH-145) Add ability to tar.gz a given template
This commit adds the `build` package, which will take a given
template path and output dir, and generate a .tar.gz of the
template path's contents.

This work is built upon the previous effort in GH-137 which
implemented the TAR and GZIP utility methods.

Some preflight sanity checks are performed on the template path
to ensure it meets the minimum requirements of a PCT:

- Are we in a valid module's root dir?
- Does the directory path exist?
- Does it contain a `pct-config.yml` at root?
- Does it contain a `content` dir at root?

If those conditions are satisfied, the template directory is
first TAR'd and then GZIP'd, and the resulting `.tar.gz` output to
the specified directory.

If any operation fails along the way (pre checks / TAR/GZ operation),
an error is returned.