(GH-144) Redirect `pct build` to PCT template build helper
Plumb the `pct build` cmd to the `pct.Build` function to allow PCT
templates to be TAR.GZ'd to a given location.

The CLI options for `pct build` are:
- `--sourcedir`: Unspecified, it takes `$CWD`, otherwise this param
points to the template you wish to TAR.GZ
- `--targetdir`: Unspecified, it will output to `$CWD/pkg`, otherwise
it will output to the path provided in this param

Acceptance tests have also been added for the `pct build` command.

A fix to `gzip.go` was required to ensure that the full diretory path
was created prior to the `os.Create` call, otherwise this would fail