(GH-144) Add `testutils.GetTmpDir` helper function
On macOS systems, the `TempDir` func in the `testing` package will
potentially return the symlink to the dir, rather than the actual
path (`/private/folders/...` vs `/var/private/folders/...`). This
can cause issues when there is a call to `os.Getwd` and you're
attempting to validate the output from the app meets the expectations
built upon the output of `testing`'s `TempDir` function.

This commit adds the `GetTmpDir` func to the `testutils` package.
The output of `t.TempDir` will be passed to `filepath.EvalSymlinks`
to resolve the symlink (if applicable) to the fully qualified path.