(GH-167) Add `pct install` CLI command
Add a new `pct install` CLI command and plumb to the `pct.Install`
function to allow PCT template archives (in TAR.GZ format) to be
installed to the default template path or the template path defined
using the `--templatepath` option.

The CLI options for `pct install` are:
- `--templatepath`: Unspecified, it takes the default template path
which is `$INSTALL_PATH/templates`

This commit also catches the underlying errors from function calls
within `pct.Install` and returns a more user friendly error to the
CLI.

Acceptance tests have also been added for the `pct install` command.