(GH-167) Refactor `cmd/install` for more testability
This commit refactors `cmd/install.go`, primarily to accept a new
type of `PctInstaller`. This allows the command creation to be
tested by passing in a new `mock.PctInstaller` and verifying the
expected args are passed to it.

This structure will act as a template for future commands and as
a reference for refactoring `build` and `new`, in future.