(GH-182) Add author to template display
Prior to this commit, template authors were not included in the display
output when running pct new --list

As templates are installed per author as namespaces, the list of
templates now also includes their author in addition to their display
name, name, and type.

Adds unit tests for the FormatTemplates public function.