(MAINT) Refactor FormatTemplates for output
Prior to this commit, FormatTemplates did not emit any output
of its own; instead, it called Printf and Render to emit
the output inside itself.

This commit ensures that the output of table formatting is
captured and returned to enable the function to be testable;
this also prepares for a future commit where the emitting of
the formatted templates is pulled out of this function to
separate concerns; Format does not imply emission and this
change allows for some simplification.