(maint) Set tmpdir to be on the same drive as the workspace
When installing a template we extract these to a temporary directory, read the config and then move them to their permanent home. This caused an issue in the windows CI as the tmp directory was on a different drive from the workspace.
This commit sets the tmp dir location to be inside the GitHub workspace.