(GH-201) Cache PCT templates during preExec of `pct new`
Prior to this commit, there were multiple calls to `pctApi.List()`
when the `cmd/new` package was being initialised.

This commit reduces this to a single call, performed in the
`PreRunE` stage of the `cobra.Command` instantiation, and caches
the result to a new variable called `cachedTemplates`.

Any subsequent calls to `pctApi.List()` now instead use the
cached data.

Also, prior to this commit, the `pctApi.GetInfo` and `pctApi.Get`
function arguments were:

```go
(templateCache string, selectedTemplate string)
```

This commit changes the method signatures of these methods to:
```go
(templateDirPath string)
```

`pctApi.GetInfo` will now need to receive the full path to the
template that you wish to retrieve info on. The returned data is
still the same.

Any calls made to these methods has been updated as part of this
change, too.

Finally, the tests within `internal/pkg/pct/pct_test.go` have been
updated to reflect the change in behaviour of the `pctApi.Get`
method.