(GH-184) Enable install of remote templates
Prior to this commit the only way to install a PCT template was to have
already downloaded the tar.gz locally.

This commit updates the `Install` function to handle a `templatePkg`
specified as a valid URL, downloading the file to a temporary folder
and then handing the file off to the remainder of the function to
process like any local file.

This commit also implements the new `downloadTemplate` function, which
handles downloading and placing the template archive file, returning the
path to it.