(GH-183) Update `pct new` to handle namespaced & versioned templates
Prior to this commit, the `pct new` command would only be capable
of processing templates in the old format (`< 0.4.0`) of
`id/(pct-config.yml|content/)`.

This commit adds support for the new namespaced and versioned PCT
template structure of `namespace/id/ver/(pct-config.yml|content/)`
which will become the standard from `0.4.0` onwards.

Also as part of this commit:
- The test resources under `internal/pkg/pct/testdata/examples`
have been updated to the new template format
- Unit and acceptance tests updated
- Re-enabled the tests temporarily backed out until GH-183 complete