(GH-183) Display more useful error for invalid template format
Prior to this commit, if an invalid template format is provided,
the "Couldn't find an installed template that matches" error
message displays, which can be slightly misleading given that it
may exist, but, the selected template was provided in an invalid
format.

This commit adds a check to the `validateArgCount` method in `new.go`
to ensure the selected template is in the `AUTHOR/ID` format and
errors if it is not.