(GH-208) Prepare repo for multi-tool building
Prior to this commit the folder layout and structure for PDKGo was only partially
ready for multi-tool builds; this commit moves all of the PCT code from cmd into
cmd/pct, moves the PCT main.go file into cmd/pct, and updates the goreleaser
config to point to the new location for main.

This will enable us to add a new binary main.go in cmd/new_project_name and update
the goreleaser to also build that tool.

There is *likely* some shared code across the tools that can be refactored and the
project at large may need more extensive refactoring in the near future.