(GH-222) Add Honeycomb/OpenTelemetry to PCT
This commit takes the first step towards implementing telemetry for the
PCT binary by adding open telemetry and the Honeycomb exporter to it;
in this initial phase, the only information being captured is the OS
and architecture of the machine running PCT.

This implementation is designed with three compile-time flags in mind:

1. Whether or not telemetry should be turned on at all
2. The Honeycomb API key to use if telemetry is turned on
3. The Honeycomb dataset to send telemetry to if turned on

These are handled in the ldflags and build tag of goreleaser and the
build scripts optionally set these values to coherent defaults (do
not report the telemetry, use a special identifier to make it obvious
the binary is not configured to send telemetry to Honeycomb).

This immediately allows us to create builds with/without telemetry via
calling goreleaser.

The implementation in only performs telemetry setup if:

1. The build tag for telemetry is set to true
2. The Honeycomb API Key does not equal `not_set`
3. The Honeycomb dataset does not equal `not_set`

If any of those conditions are not met, the binary *can not* report
any telemetry. Moreover, in local testing it turns out that if the
telemetry backends are not configured the inclusion of any tracers
and spans are wholly ignored.

The only thing that *must* be propagated is the current context.
Exactly how to do this with the commands/subcommands will need to
be determined.