(GH-223) Implement UUID for telemetry
Prior to this commit, a binary reporting telemetry would report on the Operating
System and Architecture of the machine PCT was running on; this commit adds a new
data key for UUID to enable per-machine telemetry and analysis.

The UUID itself is further hashed to our own namespace to prevent even the
possibility of leaking a UUID which could be reversed to discover the user.