(MAINT) Add debug launcher for pct new --list
This commit adds a launch.json file to the workspace and includes a
configuration for launching a debug execution of pct new --list; this
implementation builds with the telemetry tag passed and defaults to
expecting that the baker-round repo (with all the templates) is made
available in the parent folder of this project.

The value for the template path is overwritable when running the debug
launcher. In tests, configurations which do not use a particular input
do not prompt for it (so a future configuration for debugging something
like pct --version would not prompt for the template directory).

This should serve as a prototype for future debug launch configurations
and ease the contribution/development flow for the team and others.