(GH-136) Add Clone Template Functionality
Added `--git-uri` flag that takes a string variable as its argument.

Modified existing code in the `executeInstall` method so that it checks if the `--git-uri` flag is being used. The appropriate function for template installation will be called based on this check.

`preExecute` function changed to provide the correct template installation path for the `--git-uri` flag.

Implemented functions that allow for the cloning of templates from a remote repository using a git URI.

Added unit and acceptance tests where relevant.

Added appropriate documentation to `CHANGELOG.md` and `README.md`.