(MAINT) Refactor telemetry to pass root span name
Prior to this commit, the telemetry package hard coded the root span to
execution; this commit modifies the Start function to take an additional
paramater, rootCommandName, which is passed when called in main to ensure
that the root span is appropriately named. This will enable broader reuse
by other PDK tools such as PRM.

This commit also renames the variable containing the root span in main
from span to parentSpan in preparation for spinning up a child span in
main instead of in each subcommand.