(MAINT) Standardize telemetry intialization
This commit modifies the Start function of telemetry to set the Service to PDK
instead of PCT (as the root span name is now pct, allowing us to be clear that
pct is a component of the PDK) and the parent tracer to report the actual
library being used instead of clobbering the name to pct needlessly; this
will make troubleshooting for future maintainers easier as it ensures that the
library identified in honeycomb traces is the actual library used to trace the
code execution instead of being named after the command running.

The library naming was already in use for child spans, so this standardizes the
root span to use the same, more sensible convention.