(GH-258) Remove calls to Github API for packages
Prior to this commit, there was a call to api.github.com
to obtain the releases listing, which was then
subsequently processed to determine the latest ver.

Calls to the API were triggering throttling mechanisms
and causing CI test runs to fail.

This commit modifies the `.goreleaser.yml` config
to remove the version from the build artefacts published,
which then allows us to hit `releases/latest/` on
Github (bypassing the API).

The `install.(ps1|sh)` scripts have also been updated
to use this new URI.