(GH-285) Fix argumentless execution
Prior to this commit, when running PCT without arguments, the
GetCalledCommand function would fatally error on an index out
of range failure because the logic acts on the arguments and
always expects at least one.

This commit adds a guard to the GetCalledCommand function to
return empty strings for subcommand and arguments if no args
are passed at the commandline.