(GH-287) Ensure unconfigured telemetry fails cleanly
Prior to this commit, if the Honeycomb API key and/or dataset
were not passed at build time, the binary with telemetry would
first log warnings about this, then fail esoterically at the
end of an execution when attempting to shutdown an uninitialized
provider.

This commit modifies the Start function for telemetry to instead
write a single clean fatal-level error message if the required
values for telemetry have not been set, reducing confusion and
unneccessary code execution.