(GH-309) Genericise `acceptance/testutils` package
Prior to this commit, the utility functions within the `testutils`
package were working on the assumption that `pct` was the only
binary we would need to invoke for acceptance tests.

This commit genericises the utility methods by:
- Adding a `SetAppName` function
- Replacing any hardcoded string literals of "pct" and referencing
the `app` var set by `SetAppName`
- Renamed the `RunPctCommand` func to `RunAppCommand`

This change will allow us to import this package in to other
repos and use it for acceptance tests there. The first use case
will be for `puppetlabs/prm`.

Also part of this commit is the update of the tests that live in
the `acceptance` dir based on the above changes