(GH-11) Add docs package
This commit adds the docs package to the project, which contains
interfaces and functions for finding, parsing, filtering, and
rendering documentation in the terminal.

It relies on embedding the documents in the binary via the embed
library; those documents are then parsed for their frontmatter
metadata and body markdown and cached on an instance.