(GH-11) Add md package to docs
This commit extends the existing docs/md submodule (a convenience for
adding docs to the hugo site) into a properly importable module. This
module has a single purpose: to embed the documentation in any binary
it is included in and make those documents available via the embed
library.

This allows a consumer to parse and render documentation without
having to write the documents themselves to disk outside the binary.

A caveat to this approach is that documentation cannot be updated
independently of a build; however it simplifies overall deployment
and ensures the docs at release time are always included.