(GH-11) Update docs for terminal rendering
Prior to this commit, the documentation for the project included minimal
frontmatter metadata and utilized hugo shortcodes.

This commit updates the documentation to add the following frontmatter fields:

- category (concept, narrative): What sort of documentation a user can expect,
  used for filtering down available topics
- tags (string): A freeform list of tags to associate to documents, used for
  filtering down available topics
- skipTerminal (bool): Whether or not a document should *not* be rendered in
  the terminal; typically for index pages which are not themselves useful

This commit also removes all shortcodes from the documentation; while the
shortcodes are themselves extremely useful for the website display, they
break the initial parsing for the terminal. A future commit adding handlers
to the docs package for these shortcodes may enable putting them back.

A few notes on things that do not work with the terminal renderer currently:

- Quote blocks behave strangely; they do not respect paragraph semantics of
  markdown (a single newline is treated as a para break instead of continuing
  the same para) and they do not display nested code blocks at all.
- Something about tables breaks in file-parsed markdown but not in string-built
  markdown; this prematurely ends the entire file but fails silently.

These problems will need to be addressed in the future, possibly via upstream
fixes to the glamour library (which renders markdown in the terminal).