(MAINT) Handle space separated `toolArgs` param for PRM tests
Prior to this commit, any acceptance tests that made use of the
`toolArgs` flag in the `PRM` app, would fail as the args were
all split on a space delimiter prior to being passed to the app.
This would mangle the `toolArgs` parameter and cause failures.

This commit adds a function to `testutils` that will iterate over
the `cmds` slice and if `--toolArgs` is present, amalgamate it in
to one single arg.

If no `--toolArgs` parameter is present, it simply returns `cmds`
as is.

GH-52 has been raised to enhance the PRM app to handle `toolArgs`
with a value that contains spaces in it. This commit can potentially
be dropped when GH-52 is complete