(GH-342) Add config_processor package
Prior to this commit, the only implementation of an interface
for the ConfigProcessor lived in the install package; moreover,
this interface only included the ProcessConfig function.

This commit creates a new public package, config_processor, which
contains only a single interface: ConfigProcessorI.

This interface includes both the ProcessConfig function and the new
CheckConfig function. This will allow external libraries/tools  to
re-use our code so long as they implement a processor for their own
needs which can perform those tasks.

This commit does **not** implement or otherwise alter the private
configuration processor (pct_config_processor) or any other files.