(GH-342) Implement CheckConfig in private processor
Prior to this commit, functional logic for validating a config file
lived not in the private processor (pct_config_processor) but in the
private build package as a private function.

This commit adds the CheckConfig function to the private processor
as a public function so that the private processor becomes a valid
implementation of the ConfigProcessorI interface.

It slightly refactors the implementation from the build package in
this transfer, as the logic for reading a configuration file is
already defined in a private function for the private processor. It
also ensures that the full path to the configuration file being checked
is returned in any error messages; the prior implementation only
specified 'pct-config.yml' which is neither generic for use in other
libraries/tools, nor specific if the caller is pointing to a config
anywhere but in the current working directory.

This commit **does not** replace any upstream package usage to point
to the new function, it **only** makes the private processor a valid
implementation of the ConfigProcessorI interface.