(GH-342) Remove refs to PCT in build package
Prior to this commit, the private build package contained
several hard-coded references to pct-config.yml preventing
this package from becoming truly generic.

This commit removes the hard-coded config file name by
extending the Builder struct to include the ConfigFile
key for that value instead.

This commit updates tests and the build command to reflect
this refactor.