(GH-342) Refactor build.Build() for generic input
Prior to this commit the `build.Build()` function took `templatePath` as
the input for the directory to attempt to package and referenced the
"template" in both comments and messages.

This commit updates the function to instead take `sourceDir` as the
input for the directory to attempt to package and removes all references
to "template" from comments and messages, replacing it where coherent
with "project"

This further genericizes the build package for external library usage.