(GH-342) Genericize build cmd
Prior to this commit, the cobra `build` command used package variables
and instantiated a PCT-specific `build` command in the `CreateCommand()`
function. Any other cobra applications wanting to use the now-generic
functionality from the public `build` package would need to reimplement
this command wholly themselves as it hard-codes to PCT's packaging
paradigm (i.e. it references "templates", expects a `pct-config.yml`,
and instantiates the private PCT `ConfigProcessor`).

This commit genericizes the `build` command by:

1. Replacing the package variables with a `BuildCommand` struct,
   declaring a `BuildCommandI` interface, and updating the functions
   (`CreateCommand()`, `preExecute()`, and `execute()`) to be attached
   to the new `BuildCommand` struct.
2. Replacing references in the `build` command's reference doc help text
   and error/info messaging, updating default language to "project" and
   extending the `BuildCommand` struct to include defining a
   `ProjectType` to clarify those docs/messages (i.e. "Builds a package
   from the template project" for PCT where `ProjectType` is set in the
   struct as `template`).
3. Moving the instantiation of the struct to `main` where it  defines
   the `ProjectType` as `template` and instantiates the `build.Builder`
   struct, passing in the private `PctConfigProcessor`.

During testing for this change, I noticed that the unit tests for the
`build` command did not mock the `build.Builder`; this was because the
`CreateCommand()` function instantiates that struct inside itself.
Therefore, the unit tests were _actually_ trying to build packages.

This commit reworks and simplifies those tests so that they use a new
mock `Builder` instead of functionally (and without documentation)
running integration tests for the `build` package.