(GH-349) Move install logic from config_processor
Prior to this commit, the private processor included the ProcessConfig
public function which called the private functions readConfig and
setupTemplateNamespace to place a package in the right folder structure
so that it is recognized as an installed package.

This is an install function placed in the config processor.

This commit replaces the ProcessConfig function in the
config_processor.ConfigProcessorI interface with GetConfigMetadata,
which returns the new ConfigMetadata struct also newly defined in
config_processor. This enables a consumer of a ConfigProcessorI
implementation to retrieve the metadata needed to install a package
generically. The prior implementation conflated configuration processing
with install logic.

This commit adds a new public function to the install package,
InstallFromConfig. This function combines the logic from the removed
ProcessConfig and setupTemplateNamespace functions from the private
config processor, genericising it in the process.

This commit extends the install.Installer struct to take a config file
name, so that no hard coding of the config file name is needed in the
package; this makes it generically reusable regardless of config file.

Additionally, this bug corrects a minor issue in the cleanup for an
installed package, ensuring the temp folder the package is gunzipped to
can actually be deleted on Windows by replacing the `AFS.Remove()` call
with one to `AFS.RemoveAll()`, which removes a folder and its children
recursively. `AFS.Remove()` previously errored on Windows due to the
gunzipped tar file remaining after the untarred contents were moved.