(MAINT) Handle default for GOAMD64 builds
As of go 1.18, the GOAMD64 has been introduced to represent the
different microarchitecture levels to compile for when GOARCH is
amd64.

It would appear that GOAMD64 is not available in the runtime package,
so in this commit we are making an assumption that we will be building
for the default of v1. This is currently true for goreleaser config
in PRM and PCT.